Lab03S01/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
pandas==2.2.2
tqdm==4.66.5
tenacity==9.0.0
